# documentation/final-project-api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/questions/{id}": {
            "get": {
                "tags": [
                    "question-controller"
                ],
                "operationId": "getQuestionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "question-controller"
                ],
                "operationId": "updateQuestion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "question-controller"
                ],
                "operationId": "deleteQuestion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/questions": {
            "get": {
                "tags": [
                    "question-controller"
                ],
                "operationId": "getAllQuestions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Question"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "question-controller"
                ],
                "operationId": "createQuestion",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QuestionDTO": {
                "type": "object",
                "properties": {
                    "questionText": {
                        "type": "string"
                    },
                    "optionA": {
                        "type": "string"
                    },
                    "optionB": {
                        "type": "string"
                    },
                    "optionC": {
                        "type": "string"
                    },
                    "optionD": {
                        "type": "string"
                    },
                    "correctAnswer": {
                        "type": "string"
                    },
                    "quizId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "Question": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "questionText": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionA": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionB": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionC": {
                        "type": "string"
                    },
                    "optionD": {
                        "type": "string"
                    },
                    "correctAnswer": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "correctAnswer",
                    "optionA",
                    "optionB",
                    "questionText"
                ]
            },
...
        }
    }
}