# triviahub/src/main/java/com/triviahub/triviahub/model/QuestionDTO.java
package com.triviahub.triviahub.model;

/**
 * Data Transfer Object for creating and updating Question entities.
 * This class is used to transfer data between the client and the server, 
 * without exposing the internal database entities.
 */
public class QuestionDTO {
    private String questionText;
    private String optionA;
    private String optionB;
    private String optionC;
    private String optionD;
    private String correctAnswer;

    /**
     * The ID of the Quiz this question should be associated with.
     * This replaces the direct userId link, as questions now belong to quizzes.
     */
    private Long quizId;

    //Getters and Setters

    public String getQuestionText() { return questionText; }
    public void setQuestionText(String questionText) { this.questionText = questionText; }

    public String getOptionA() { return optionA; }
    public void setOptionA(String optionA) { this.optionA = optionA; }

    public String getOptionB() { return optionB; }
    public void setOptionB(String optionB) { this.optionB = optionB; }

    public String getOptionC() { return optionC; }
    public void setOptionC(String optionC) { this.optionC = optionC; }

    public String getOptionD() { return optionD; }
    public void setOptionD(String optionD) { this.optionD = optionD; }

    public String getCorrectAnswer() { return correctAnswer; }
    public void setCorrectAnswer(String correctAnswer) { this.correctAnswer = correctAnswer; }

    public Long getQuizId() { return quizId; }
    public void setQuizId(Long quizId) { this.quizId = quizId; }
}